test(runtime): isolate catch predicate fallthrough regressions
