+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 - 2026-08-22: Continued #2564 test-module refactoring by moving the focused
   `THROW oException` chaining and `Exception.UserValue` reference regression
-  into `test_prg_engine_control_flow_throw_exception_chain.cpp`. The existing
+  into `test_prg_engine_control_flow_throw_exception_chain.cpp`, and the
+  false-`CATCH WHEN` predicate fallthrough regressions into
+  `test_prg_engine_control_flow_catch_predicate_fallthrough.cpp`. The existing
   `test_prg_engine_control_flow` executable, invocation order, assertions, and
   runtime behavior remain unchanged.
